sevila/requirements.txt
contexttimer
decord~=0.6.0
einops>=0.4.1
fairscale==0.4.4
ftfy~=6.2.3
iopath~=0.1.10
ipython
omegaconf~=2.3.0
opencv-python-headless==4.5.5.64
opendatasets~=0.1.22
packaging
pandas~=2.2.2
plotly~=5.23.0
pre-commit
pycocoevalcap~=1.2
pycocotools~=2.0.8
python-magic~=0.4.27
scikit-image~=0.24.0
sentencepiece
spacy~=3.7.5
streamlit~=1.37.1
timm==0.4.12
torch>=1.10.0
torchvision~=0.19.0
tqdm~=4.66.5
transformers>=4.25.0
webdataset~=0.2.96
wheel

pillow~=10.4.0
numpy~=1.26.4
matplotlib~=3.9.2
scipy~=1.14.0
requests~=2.32.3
PyYAML~=6.0.2
regex~=2024.7.24
gradio~=3.30.0
setuptools~=72.1.0